Fixes 'implicit conversion' warnings in DDGHistoryViewController

+++ DuckDuckGo/DDGHistoryViewController.m
@@ -384,7 +384,7 @@
         DDGHistoryItem *historyItem = [self.fetchedResultsController objectAtIndexPath:historyIndexPath];
         //    [self.historyProvider relogHistoryItem:historyItem];
         DDGStory *story = historyItem.story;
-        int readabilityMode = [[NSUserDefaults standardUserDefaults] integerForKey:DDGSettingStoriesReadabilityMode];
+        NSInteger readabilityMode = [[NSUserDefaults standardUserDefaults] integerForKey:DDGSettingStoriesReadabilityMode];
         if (nil != story)
             [self.searchHandler loadStory:story readabilityMode:(readabilityMode == DDGReadabilityModeOnExclusive || readabilityMode == DDGReadabilityModeOnIfAvailable)];
         else
